Non-templated individual book view
