+++ frontend/src/router/index.ts
@@ -5,6 +5,7 @@ import FAQView from '@/views/FAQView.vue'
 import IndividualStoryView from '@/views/IndividualStoryView.vue'
 import AboutView from '@/views/AboutView.vue'
 import ResourcesView from '@/views/ResourcesView.vue'
+import DisorderedCosmosView from '@/views/reads/DisorderedCosmosView.vue'
 import ContactView from '@/views/ContactView.vue'
 import DataDashboardDatasetView from '@/views/DataDashboardDatasetView.vue'
 import DataDashboardView from '@/views/DataDashboardView.vue'
@@ -47,6 +48,11 @@ const router = createRouter({
       component: ResourcesView
     },
     {
+      path: '/book_01',
+      name: 'book_01',
+      component: DisorderedCosmosView
+    },
+    {
       path: '/contact',
       name: 'contact',
       component: ContactView
@@ -62,6 +68,7 @@ const router = createRouter({
     },
     {
       path: '/data',
+      name: 'data',
       redirect: '/data-dashboard'
     },
     {
